Throw errors for global handling and restructure socket files
