+++ src/index.ts
@@ -4,6 +4,7 @@ import { Server } from 'socket.io';
 import { generalLogger } from './utils/loggers';
 import { ServerToClientEvents, SocketWithAuth } from './sockets/socket.types';
 import { createTokenMiddleware } from './sockets/socket.middlewares';
+import { socketErrorHandler } from './sockets/socketErrorHandler';
 import onConnection from './sockets';
 import app from './app';
 
@@ -12,7 +13,10 @@ const server = createServer(app);
 const io = new Server<ServerToClientEvents, ServerToClientEvents, DefaultEventsMap, SocketWithAuth>(server);
 
 // TODO - add a validation middleware
-io.of('polls').use(createTokenMiddleware).on('connection', onConnection)
+// use a try/catch block inside of 'connection' to catch the errors
+io.of('polls').use(createTokenMiddleware).on('connection', onConnection);
+
+io.of('polls').use(socketErrorHandler);
 
 const port = process.env.PORT || 3000;
