+++ src/sockets/socket.types.ts
@@ -3,7 +3,7 @@ import { RejoinPollFields } from "../api/polls/poll.types";
 export interface ServerToClientEvents {
     disconnect: () => void;
     hello: (message: string) => void;
-    error: (error:string) => void;
+    error: (error: string) => void;
 }
 
 export interface SocketWithAuth {
